Save movzwl instructions

+++ drivers/staging/snappy/csnappy_decompress.c
@@ -106,7 +106,7 @@ static const uint16_t char_table[256] = {
  * Note that this does not match the semantics of either memcpy()
  * or memmove().
  */
-static inline void IncrementalCopy(const char *src, char *op, int len)
+static inline void IncrementalCopy(const char *src, char *op, ssize_t len)
 {
 	DCHECK_GT(len, 0);
 	do {
@@ -147,7 +147,7 @@ static inline void IncrementalCopy(const char *src, char *op, int len)
  * position 1. Thus, ten excess bytes.
  */
 static const int kMaxIncrementCopyOverflow = 10;
-static inline void IncrementalCopyFastPath(const char *src, char *op, int len)
+static inline void IncrementalCopyFastPath(const char *src, char *op, ssize_t len)
 {
 	while (op - src < 8) {
 		UNALIGNED_STORE64(op, UNALIGNED_LOAD64(src));
